requirements.txt
# Core Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-Login==0.6.3

# Environment & Configuration
python-dotenv==1.1.0

# Data Processing
pandas==2.2.3

# Database
SQLAlchemy==2.0.40

# HTTP & API
requests==2.32.3

# Utilities
gunicorn==23.0.0
email-validator==2.2.0
